Getting newest changes

+++ BlazorApp/DataStreaming/Data.cs
@@ -7,10 +7,10 @@ namespace BlazorApp.DataStreaming
     public class Data
     {
         public string ReportType { get; set; }
-        public long NumericValue { get; set; }
+        public int NumericValue { get; set; }
         public DateTime LogTime { get; set; }
 
-        public Data(string reportType, long numericValue, DateTime logTime)
+        public Data(string reportType, int numericValue, DateTime logTime)
         {
             ReportType = reportType;
             NumericValue = numericValue;
@@ -20,7 +20,7 @@ namespace BlazorApp.DataStreaming
         public Data(SqlDataReader reader)
         {
             ReportType = (string) reader[0];
-            NumericValue = (long) reader[1];
+            NumericValue = (int) reader[1];
             LogTime = (DateTime) reader[2];
         }
